tts-server/requirements.txt
flask>=3.1.3
requests>=2.32.0
python-dotenv>=1.0.0
elevenlabs>=1.0.0
